Run e2e-test via main
