+++ Makefile
@@ -46,12 +46,12 @@ build-e2e: ## Build end-to-end test binary
 	$(DOCKER_CMD) go build $(GOGCFLAGS) -o bin/e2e github.com/openshift/machine-api-operator/tests/e2e
 
 test-e2e:
-	go test -v ./test/e2e/...
+	go run ./test/e2e/*.go -alsologtostderr
 
 .PHONY: test
 test: ## Run tests
 	@echo -e "\033[32mTesting...\033[0m"
-	$(DOCKER_CMD) go test ./pkg... ./cmd/...
+	$(DOCKER_CMD) go test ./...
 
 .PHONY: image
 image: ## Build docker image
